requirement.txt
streamlit==1.32.0
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
protobuf==3.20.3